artgram-microservices/services/auth-service/requirements.txt
Django==5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
PyJWT==2.8.0
requests==2.31.0
django-cors-headers==4.3.1
celery==5.3.4
redis==5.0.1
python-consul==1.1.0
Pillow
